pyproject.toml
[tool.poetry]
name = "aia-utils"
version = "0.1.0"
description = ""
authors = ["Edgar Sánchez <edgar.sanchez@mercadolibre.cl>"]
packages = [{include = "src"}]

[tool.pytest.ini_options]
pythonpath = "src"

[tool.poetry.dependencies]
python = "^3.11"
confluent-kafka = "^2.3.0"
coverage = "^7.3.2"
PyYAML = "^6.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
pytest-dotenv = "^0.5.2"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
